refactor: relocate page components
