+++ src/components/useLiveDevices.js
@@ -1,6 +1,6 @@
 import {useCallback, useMemo, useState} from "react";
-import {filterNoise, normalizeSensorData} from "../../utils";
-import {useStomp} from "../../hooks/useStomp";
+import {filterNoise, normalizeSensorData} from "../utils";
+import {useStomp} from "../hooks/useStomp";
 import {SENSOR_TOPIC} from "./dashboard.constants";
 
 function mergeControllers(a = [], b = []) {
