+++ src/pages/Dashboard/components/DeviceCard.jsx
