+++ src/pages/Dashboard/components/DeviceCard.module.css
