+++ src/pages/Live/components/Live.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import SpectrumBarChart from "../SpectrumBarChart";
+import SpectrumBarChart from "../../../components/SpectrumBarChart";
 import TopicSection from "./TopicSection";
 import NotesBlock from "./NotesBlock";
-import styles from "../SensorDashboard/SensorDashboard.module.css";
+import styles from "./SensorDashboard/SensorDashboard.module.css";
 
 function Live({
     filteredSystemTopics = {},
