+++ src/pages/Reports/components/ReportControls.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../SensorDashboard/SensorDashboard.module.css';
+import styles from '../../Live/components/SensorDashboard/SensorDashboard.module.css';
 
 function ReportControls({
   fromDate,
