+++ tests/SensorDashboardOverview.test.jsx
@@ -3,7 +3,7 @@ import { render, screen, within, act } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
 
-import SensorDashboard from '../src/components/SensorDashboard';
+import SensorDashboard from '../src/pages/Live/components/SensorDashboard';
 import { FiltersProvider } from '../src/context/FiltersContext';
 
 // Mock WebSocket hook so we can control messages
@@ -21,8 +21,8 @@ vi.mock('../src/hooks/useStomp', () => ({
 // Stub components that are not relevant for this test
 vi.mock('../src/components/SpectrumBarChart', () => ({ default: () => <div /> }));
 vi.mock('../src/components/Header', () => ({ default: () => <div /> }));
-vi.mock('../src/components/dashboard/TopicSection', () => ({ default: () => <div /> }));
-vi.mock('../src/components/dashboard/NotesBlock', () => ({ default: () => <div /> }));
+vi.mock('../src/pages/Live/components/TopicSection', () => ({ default: () => <div /> }));
+vi.mock('../src/pages/Live/components/NotesBlock', () => ({ default: () => <div /> }));
 
 test('overview items reflect live_now data', () => {
   render(
